+++ src/renderer/components/dialogs/MailtoDialog.tsx
@@ -9,7 +9,7 @@ import { DialogProps } from './DialogController'
 
 import { C } from 'deltachat-node/dist/constants'
 import { ChatListPart, useLogicVirtualChatList } from '../chat/ChatList'
-import { AutoSizer } from 'react-virtualized'
+import AutoSizer from 'react-virtualized-auto-sizer'
 import { useChatList } from '../chat/ChatListHelpers'
 import { selectChat } from '../../stores/chat'
 
@@ -68,11 +68,12 @@ export default function MailtoDialog(props: {
                     rowCount={chatListIds.length}
                     width={width}
                     height={height}
+                    itemKey={index => 'key' + chatListIds[index]}
                   >
-                    {({ index, key, style }) => {
+                    {({ index, style }) => {
                       const [chatId] = chatListIds[index]
                       return (
-                        <div style={style} key={key}>
+                        <div style={style}>
                           <ChatListItem
                             chatListItem={chatCache[chatId] || undefined}
                             onClick={onChatClick.bind(null, chatId)}