+++ src/renderer/components/dialogs/ViewProfile.tsx
@@ -16,7 +16,7 @@ import { C } from 'deltachat-node/dist/constants'
 import { ScreenContext } from '../../contexts'
 import { Avatar } from '../Avatar'
 import { useLogicVirtualChatList, ChatListPart } from '../chat/ChatList'
-import { AutoSizer } from 'react-virtualized'
+import AutoSizer from 'react-virtualized-auto-sizer'
 import MessageBody from '../message/MessageBody'
 
 const ProfileInfoName = ({
@@ -178,11 +178,12 @@ export default function ViewProfile(props: {
                     rowCount={chatListIds.length}
                     width={width}
                     height={height}
+                    itemKey={index => 'key' + chatListIds[index]}
                   >
-                    {({ index, key, style }) => {
+                    {({ index, style }) => {
                       const [chatId] = chatListIds[index]
                       return (
-                        <div style={style} key={key}>
+                        <div style={style}>
                           <ChatListItem
                             chatListItem={chatCache[chatId] || undefined}
                             onClick={onChatClick.bind(null, chatId)}